포지션 릴레이 소수점 두자리까지 하도록 수정

+++ Assets/MirageTest/Scripts/Messages/PositionRelayMessage.cs
@@ -7,7 +7,7 @@ namespace MirageTest.Scripts.Messages
     public class PositionRelayMessage
     {
         public uint netId;
-        public float positionX;
-        public float positionY;
+        public short positionX;
+        public short positionY;
     }
 }
